+++ src/main/java/com/wedevol/springdatacouchbase/core/controller/PlaceController.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import com.wedevol.springdatacouchbase.core.dao.doc.PlaceDoc;
@@ -32,7 +31,6 @@ public class PlaceController {
 
   @RequestMapping(value = "/{placeId}", method = RequestMethod.GET)
   @ResponseStatus(HttpStatus.OK)
-  @ResponseBody
   public PlaceDoc findById(@PathVariable Long placeId) {
     logger.info("Find place by id: {}", placeId);
     return placeService.findById(placeId);
@@ -40,7 +38,6 @@ public class PlaceController {
 
   @RequestMapping(value = "", method = RequestMethod.POST)
   @ResponseStatus(HttpStatus.CREATED)
-  @ResponseBody
   public PlaceDoc create(@Valid @RequestBody PlaceDoc place) {
     logger.info("Create place");
     return placeService.create(place);